requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
names==0.3.0
paho-mqtt==1.6.1
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.3
websockets==10.1
